split utility tests
