+++ QualityControl/test/lib/utils/httpRequests.test.js
@@ -12,56 +12,12 @@
  * or submit itself to any jurisdiction.
  */
 
-import { stub } from 'sinon';
 import nock from 'nock';
-import { ok, strictEqual, deepStrictEqual, rejects } from 'node:assert';
-import { suite, test, beforeEach, before } from 'node:test';
-import { errorHandler } from '../../../lib/utils/errorHandler.js';
+import { strictEqual, deepStrictEqual, rejects } from 'node:assert';
+import { suite, test, before } from 'node:test';
 import { httpGetJson, httpHeadJson } from '../../../lib/utils/httpRequests.js';
 
-export const utilsTestSuite = async () => {
-  suite('Check errors are handled and sent successfully', () => {
-    let res = null;
-    beforeEach(() => {
-      res = {
-        status: stub().returnsThis(),
-        send: stub(),
-      };
-    });
-
-    test('should successfully respond with built error message when there is a message and no status', () => {
-      errorHandler('Error', 'Error', res);
-      ok(res.status.calledOnce);
-    });
-
-    test('should successfully respond with built error message and status > 500', () => {
-      errorHandler('Error', 'Error', res, 502);
-      ok(res.status.calledWith(502));
-    });
-
-    test('should successfully respond with built error message and status < 500', () => {
-      errorHandler('Error', 'Error', res, 404);
-      ok(res.status.calledWith(404));
-    });
-
-    test('should successfully respond with built error.message and status', () => {
-      const err = {
-        message: 'Test Error',
-        stack: 'Some Stack',
-      };
-      errorHandler(err, 'Error To Send', res, 502);
-      ok(res.status.calledWith(502));
-      ok(res.send.calledWith({ message: 'Error To Send' }));
-    });
-
-    test('should successfully respond with built error.message, no stack and status', () => {
-      const err = 'Test Error';
-      errorHandler(err, 'Error To Send', res, 404);
-      ok(res.status.calledWith(404));
-      ok(res.send.calledWith({ message: 'Error To Send' }));
-    });
-  });
-
+export const httpRequestsTestSuite = async () => {
   suite('"httpHeadJson" test suite ', () => {
     before(() => nock.cleanAll());
 
@@ -85,7 +41,7 @@ export const utilsTestSuite = async () => {
         .reply(200);
 
       const { status, headers } =
-        await httpHeadJson('ccdb', '8500', '/qc/some/test/123455432', { headers: { Accept: 'text' } });
+                await httpHeadJson('ccdb', '8500', '/qc/some/test/123455432', { headers: { Accept: 'text' } });
       strictEqual(status, 200);
       deepStrictEqual(headers, { lastmodified: '123132132', location: '/download/some-id' });
     });
